requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
amqp==5.3.1
    # via kombu
argon2-cffi==23.1.0
    # via minio
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
asgiref==3.8.1
    # via
    #   django
    #   django-cors-headers
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
billiard==4.2.1
    # via celery
celery==5.5.1
    # via
    #   -r requirements.in
    #   django-celery-beat
certifi==2025.1.31
    # via
    #   minio
    #   requests
    #   sentry-sdk
cffi==1.17.1
    # via argon2-cffi-bindings
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
colorama==0.4.6
    # via
    #   click
    #   pytest
cron-descriptor==1.4.5
    # via django-celery-beat
django==4.2.20
    # via
    #   -r requirements.in
    #   django-celery-beat
    #   django-cors-headers
    #   django-csp
    #   django-filter
    #   django-minio-storage
    #   django-redis
    #   django-timezone-field
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-spectacular
django-celery-beat==2.8.0
    # via -r requirements.in
django-cors-headers==4.7.0
    # via -r requirements.in
django-csp==4.0
    # via -r requirements.in
django-filter==25.1
    # via -r requirements.in
django-minio-storage==0.5.7
    # via -r requirements.in
django-redis==5.4.0
    # via -r requirements.in
django-timezone-field==7.1
    # via django-celery-beat
djangorestframework==3.16.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   drf-spectacular
djangorestframework-simplejwt==5.5.0
    # via -r requirements.in
drf-spectacular==0.28.0
    # via -r requirements.in
gunicorn==23.0.0
    # via -r requirements.in
idna==3.10
    # via requests
inflection==0.5.1
    # via drf-spectacular
iniconfig==2.1.0
    # via pytest
jsonschema==4.23.0
    # via drf-spectacular
jsonschema-specifications==2024.10.1
    # via jsonschema
kavenegar==1.1.2
    # via -r requirements.in
kombu==5.5.3
    # via celery
minio==7.2.15
    # via django-minio-storage
packaging==24.2
    # via
    #   django-csp
    #   gunicorn
    #   pytest
pillow==11.2.1
    # via -r requirements.in
pluggy==1.5.0
    # via pytest
prompt-toolkit==3.0.51
    # via click-repl
psycopg2-binary==2.9.10
    # via -r requirements.in
pycparser==2.22
    # via cffi
pycryptodome==3.22.0
    # via minio
pyjwt==2.9.0
    # via djangorestframework-simplejwt
pytest==8.3.5
    # via
    #   -r requirements.in
    #   pytest-django
pytest-django==4.11.1
    # via -r requirements.in
python-crontab==3.2.0
    # via django-celery-beat
python-dateutil==2.9.0.post0
    # via
    #   celery
    #   python-crontab
python-decouple==3.8
    # via -r requirements.in
pyyaml==6.0.2
    # via drf-spectacular
redis==5.2.1
    # via django-redis
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via kavenegar
rpds-py==0.24.0
    # via
    #   jsonschema
    #   referencing
sentry-sdk==2.26.1
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
sqlparse==0.5.3
    # via django
typing-extensions==4.13.2
    # via
    #   minio
    #   referencing
tzdata==2025.2
    # via
    #   django
    #   django-celery-beat
    #   kombu
uritemplate==4.1.1
    # via drf-spectacular
urllib3==2.4.0
    # via
    #   minio
    #   requests
    #   sentry-sdk
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
whitenoise==6.9.0
    # via -r requirements.in
